+++ cmd/cmd_renew_test.go
@@ -110,7 +110,6 @@ func Test_needRenewal(t *testing.T) {
 	for _, test := range testCases {
 		test := test
 		t.Run(test.desc, func(t *testing.T) {
-
 			actual := needRenewal(test.x509Cert, "foo.com", test.days)
 
 			assert.Equal(t, test.expected, actual)