+++ cmd/zz_gen_cmd_dnshelp.go
@@ -17,6 +17,7 @@ func allDNSCodes() string {
 		"acme-dns",
 		"alidns",
 		"auroradns",
+		"autodns",
 		"azure",
 		"bindman",
 		"bluecat",
@@ -84,7 +85,6 @@ func displayDNSHelp(name string) error {
 	ew := &errWriter{w: w}
 
 	switch name {
-
 	case "acme-dns":
 		// generated from: providers/dns/acmedns/acmedns.toml
 		ew.writeln(`Configuration for Joohoi's ACME-DNS.`)
@@ -141,6 +141,29 @@ func displayDNSHelp(name string) error {
 
 		ew.writeln()
 		ew.writeln(`More information: https://go-acme.github.io/lego/dns/auroradns`)
+
+	case "autodns":
+		// generated from: providers/dns/autodns/autodns.toml
+		ew.writeln(`Configuration for Autodns.`)
+		ew.writeln(`Code:	'autodns'`)
+		ew.writeln(`Since:	'v3.2.0'`)
+		ew.writeln()
+
+		ew.writeln(`Credentials:`)
+		ew.writeln(`	- "AUTODNS_API_PASSWORD":	User Password`)
+		ew.writeln(`	- "AUTODNS_API_USER":	Username`)
+		ew.writeln()
+
+		ew.writeln(`Additional Configuration:`)
+		ew.writeln(`	- "AUTODNS_CONTEXT":	API context (4 for production, 1 for testing. Defaults to 4)`)
+		ew.writeln(`	- "AUTODNS_ENDPOINT":	API endpoint URL, defaults to https://api.autodns.com/v1/`)
+		ew.writeln(`	- "AUTODNS_HTTP_TIMEOUT":	API request timeout, defaults to 30 seconds`)
+		ew.writeln(`	- "AUTODNS_POLLING_INTERVAL":	Time between DNS propagation check`)
+		ew.writeln(`	- "AUTODNS_PROPAGATION_TIMEOUT":	Maximum waiting time for DNS propagation`)
+		ew.writeln(`	- "AUTODNS_TTL":	The TTL of the TXT record used for the DNS challenge`)
+
+		ew.writeln()
+		ew.writeln(`More information: https://go-acme.github.io/lego/dns/autodns`)
 
 	case "azure":
 		// generated from: providers/dns/azure/azure.toml
@@ -220,10 +243,12 @@ func displayDNSHelp(name string) error {
 		ew.writeln(`Credentials:`)
 		ew.writeln(`	- "CF_API_EMAIL":	Account email`)
 		ew.writeln(`	- "CF_API_KEY":	API key`)
-		ew.writeln(`	- "CF_API_TOKEN":	API token`)
+		ew.writeln(`	- "CF_DNS_API_TOKEN":	API token with DNS:Edit permission (since v3.1.0)`)
+		ew.writeln(`	- "CF_ZONE_API_TOKEN":	API token with Zone:Read permission (since v3.1.0)`)
 		ew.writeln(`	- "CLOUDFLARE_API_KEY":	Alias to CF_API_KEY`)
-		ew.writeln(`	- "CLOUDFLARE_API_TOKEN":	Alias to CF_API_TOKEN`)
+		ew.writeln(`	- "CLOUDFLARE_DNS_API_TOKEN":	Alias to CF_DNS_API_TOKEN`)
 		ew.writeln(`	- "CLOUDFLARE_EMAIL":	Alias to CF_API_EMAIL`)
+		ew.writeln(`	- "CLOUDFLARE_ZONE_API_TOKEN":	Alias to CF_ZONE_API_TOKEN`)
 		ew.writeln()
 
 		ew.writeln(`Additional Configuration:`)
@@ -833,6 +858,7 @@ func displayDNSHelp(name string) error {
 		ew.writeln(`Additional Configuration:`)
 		ew.writeln(`	- "LINODE_HTTP_TIMEOUT":	API request timeout`)
 		ew.writeln(`	- "LINODE_POLLING_INTERVAL":	Time between DNS propagation check`)
+		ew.writeln(`	- "LINODE_PROPAGATION_TIMEOUT":	Maximum waiting time for DNS propagation`)
 		ew.writeln(`	- "LINODE_TTL":	The TTL of the TXT record used for the DNS challenge`)
 
 		ew.writeln()