+++ providers/dns/cloudflare/cloudflare_test.go
@@ -12,7 +12,8 @@ import (
 var envTest = tester.NewEnvTest(
 	"CLOUDFLARE_EMAIL",
 	"CLOUDFLARE_API_KEY",
-	"CLOUDFLARE_API_TOKEN").
+	"CLOUDFLARE_DNS_API_TOKEN",
+	"CLOUDFLARE_ZONE_API_TOKEN").
 	WithDomain("CLOUDFLARE_DOMAIN")
 
 func TestNewDNSProvider(t *testing.T) {
@@ -31,17 +32,24 @@ func TestNewDNSProvider(t *testing.T) {
 		{
 			desc: "success API token",
 			envVars: map[string]string{
-				"CLOUDFLARE_API_TOKEN": "012345abcdef",
+				"CLOUDFLARE_DNS_API_TOKEN": "012345abcdef",
+			},
+		},
+		{
+			desc: "success separate API tokens",
+			envVars: map[string]string{
+				"CLOUDFLARE_DNS_API_TOKEN":  "012345abcdef",
+				"CLOUDFLARE_ZONE_API_TOKEN": "abcdef012345",
 			},
 		},
 		{
 			desc: "missing credentials",
 			envVars: map[string]string{
-				"CLOUDFLARE_EMAIL":     "",
-				"CLOUDFLARE_API_KEY":   "",
-				"CLOUDFLARE_API_TOKEN": "",
+				"CLOUDFLARE_EMAIL":         "",
+				"CLOUDFLARE_API_KEY":       "",
+				"CLOUDFLARE_DNS_API_TOKEN": "",
 			},
-			expected: "cloudflare: some credentials information are missing: CLOUDFLARE_EMAIL,CLOUDFLARE_API_KEY or some credentials information are missing: CLOUDFLARE_API_TOKEN",
+			expected: "cloudflare: some credentials information are missing: CLOUDFLARE_EMAIL,CLOUDFLARE_API_KEY or some credentials information are missing: CLOUDFLARE_DNS_API_TOKEN,CLOUDFLARE_ZONE_API_TOKEN",
 		},
 		{
 			desc: "missing email",
@@ -49,7 +57,7 @@ func TestNewDNSProvider(t *testing.T) {
 				"CLOUDFLARE_EMAIL":   "",
 				"CLOUDFLARE_API_KEY": "key",
 			},
-			expected: "cloudflare: some credentials information are missing: CLOUDFLARE_EMAIL or some credentials information are missing: CLOUDFLARE_API_TOKEN",
+			expected: "cloudflare: some credentials information are missing: CLOUDFLARE_EMAIL or some credentials information are missing: CLOUDFLARE_DNS_API_TOKEN,CLOUDFLARE_ZONE_API_TOKEN",
 		},
 		{
 			desc: "missing api key",
@@ -57,7 +65,7 @@ func TestNewDNSProvider(t *testing.T) {
 				"CLOUDFLARE_EMAIL":   "awesome@possum.com",
 				"CLOUDFLARE_API_KEY": "",
 			},
-			expected: "cloudflare: some credentials information are missing: CLOUDFLARE_API_KEY or some credentials information are missing: CLOUDFLARE_API_TOKEN",
+			expected: "cloudflare: some credentials information are missing: CLOUDFLARE_API_KEY or some credentials information are missing: CLOUDFLARE_DNS_API_TOKEN,CLOUDFLARE_ZONE_API_TOKEN",
 		},
 	}
 
@@ -77,6 +85,116 @@ func TestNewDNSProvider(t *testing.T) {
 				assert.NotNil(t, p.client)
 			} else {
 				require.EqualError(t, err, test.expected)
+			}
+		})
+	}
+}
+
+func TestNewDNSProviderWithToken(t *testing.T) {
+	type expected struct {
+		dnsToken   string
+		zoneToken  string
+		sameClient bool
+		error      string
+	}
+
+	testCases := []struct {
+		desc string
+
+		// test input
+		envVars map[string]string
+
+		// expectations
+		expected expected
+	}{
+		{
+			desc: "same client when zone token is missing",
+			envVars: map[string]string{
+				"CLOUDFLARE_DNS_API_TOKEN": "123",
+			},
+			expected: expected{
+				dnsToken:   "123",
+				zoneToken:  "123",
+				sameClient: true,
+			},
+		},
+		{
+			desc: "same client when zone token equals dns token",
+			envVars: map[string]string{
+				"CLOUDFLARE_DNS_API_TOKEN":  "123",
+				"CLOUDFLARE_ZONE_API_TOKEN": "123",
+			},
+			expected: expected{
+				dnsToken:   "123",
+				zoneToken:  "123",
+				sameClient: true,
+			},
+		},
+		{
+			desc: "failure when only zone api given",
+			envVars: map[string]string{
+				"CLOUDFLARE_ZONE_API_TOKEN": "123",
+			},
+			expected: expected{
+				error: "cloudflare: some credentials information are missing: CLOUDFLARE_EMAIL,CLOUDFLARE_API_KEY or some credentials information are missing: CLOUDFLARE_DNS_API_TOKEN",
+			},
+		},
+		{
+			desc: "different clients when zone and dns token differ",
+			envVars: map[string]string{
+				"CLOUDFLARE_DNS_API_TOKEN":  "123",
+				"CLOUDFLARE_ZONE_API_TOKEN": "abc",
+			},
+			expected: expected{
+				dnsToken:   "123",
+				zoneToken:  "abc",
+				sameClient: false,
+			},
+		},
+		{
+			desc: "aliases work as expected", // CLOUDFLARE_* takes precedence over CF_*
+			envVars: map[string]string{
+				"CLOUDFLARE_DNS_API_TOKEN":  "123",
+				"CF_DNS_API_TOKEN":          "456",
+				"CLOUDFLARE_ZONE_API_TOKEN": "abc",
+				"CF_ZONE_API_TOKEN":         "def",
+			},
+			expected: expected{
+				dnsToken:   "123",
+				zoneToken:  "abc",
+				sameClient: false,
+			},
+		},
+	}
+
+	defer envTest.RestoreEnv()
+	localEnvTest := tester.NewEnvTest(
+		"CLOUDFLARE_DNS_API_TOKEN", "CF_DNS_API_TOKEN",
+		"CLOUDFLARE_ZONE_API_TOKEN", "CF_ZONE_API_TOKEN",
+	).WithDomain("CLOUDFLARE_DOMAIN")
+	envTest.ClearEnv()
+
+	for _, test := range testCases {
+		t.Run(test.desc, func(t *testing.T) {
+			defer localEnvTest.RestoreEnv()
+			localEnvTest.ClearEnv()
+			localEnvTest.Apply(test.envVars)
+
+			p, err := NewDNSProvider()
+
+			if test.expected.error != "" {
+				require.EqualError(t, err, test.expected.error)
+				return
+			}
+
+			require.NoError(t, err)
+			require.NotNil(t, p)
+			assert.Equal(t, test.expected.dnsToken, p.config.AuthToken)
+			assert.Equal(t, test.expected.zoneToken, p.config.ZoneToken)
+			if test.expected.sameClient {
+				assert.Equal(t, p.client.clientRead, p.client.clientEdit)
+			} else {
+				assert.NotEqual(t, p.client.clientRead, p.client.clientEdit)
 			}
 		})
 	}
@@ -107,22 +225,22 @@ func TestNewDNSProviderConfig(t *testing.T) {
 		},
 		{
 			desc:     "missing credentials",
-			expected: "invalid credentials: key & email must not be empty",
+			expected: "cloudflare: invalid credentials: key & email must not be empty",
 		},
 		{
 			desc:     "missing email",
 			authKey:  "123",
-			expected: "invalid credentials: key & email must not be empty",
+			expected: "cloudflare: invalid credentials: key & email must not be empty",
 		},
 		{
 			desc:      "missing api key",
 			authEmail: "test@example.com",
-			expected:  "invalid credentials: key & email must not be empty",
+			expected:  "cloudflare: invalid credentials: key & email must not be empty",
 		},
 		{
 			desc:      "missing api token, fallback to api key/email",
 			authToken: "",
-			expected:  "invalid credentials: key & email must not be empty",
+			expected:  "cloudflare: invalid credentials: key & email must not be empty",
 		},
 	}